refactor: restructured and refactored code
