+++ .gitignore
@@ -1,3 +1,4 @@
 .envrc
 .DS_Store
 .pytest_cache
+migrations
