+++ SocialMediaApplication/models/__init__.py
@@ -1,0 +1,7 @@
+from .user import *
+from .token import *
+from .post import *
+from .connection import *
+from .notification import *
+from .profile import *
+from .post_type import *
