+++ SocialMediaApplication/serializers/notification.py
@@ -1,0 +1,16 @@
+from .. import ma
+from SocialMediaApplication.models import User
+from marshmallow import fields
+
+class NotificationSchema(ma.Schema):
+    user = fields.Method("get_user")
+
+    def get_user(self, obj):
+        return User.query.filter_by(id=obj.user).first().username
+
+    class Meta:
+        fields = ["id", "user", "msg", "read", "published_at"]
+
+
+notification_schema = NotificationSchema()
+notifications_schema = NotificationSchema(many=True)
