+++ SocialMediaApplication/serializers/post_type.py
@@ -1,0 +1,8 @@
+from .. import ma
+
+class PostTypeSchema(ma.Schema):
+    class Meta:
+        fields = ("name",)
+
+post_type_schema = PostTypeSchema()
+post_types_schema = PostTypeSchema(many=True)
