+++ SocialMediaApplication/views/__init__.py
@@ -1,0 +1,9 @@
+from . import user
+from . import post
+from . import search
+from . import follow_request
+from . import notification
+from . import connection
+from . import post_type
+from . import auth
+from . import feed
