+++ SocialMediaApplication/views/follow_request.py
@@ -1,0 +1,161 @@
+from flask import request, jsonify, make_response
+from flask import current_app as app
+from ..models import db, User, Connection, Notification
+from ..serializers import connection_schema,connections_schema
+from ..helpers.permissions import authenticate_user
+
+
+@app.route("/follow-requests", methods=["GET"])
+@authenticate_user
+def get_my_follow_requests(**kwargs):
+    try:
+        user = kwargs.get("current_user")
+        if user:
+            follow_requests = Connection.query.filter_by(
+                receiver=user.id, accepted=False
+            ).all()
+            if len(follow_requests) == 0:
+                response_object = {
+                    "status": "success",
+                    "message": "no requests for {}".format(user.username),
+                }
+                return make_response(jsonify(response_object)), 200
+            follow_requests = connections_schema.dump(follow_requests)
+            for follow_request in follow_requests:
+                follow_request.pop("receiver")
+                follow_request["user"] = follow_request.pop("sender")
+            response_object = {"status": "success", "message": follow_requests}
+            return make_response(jsonify(response_object)), 200
+
+        else:
+            response_object = {"status": "Fail", "message": "user not found"}
+            return make_response(jsonify(response_object)), 400
+    except Exception as e:
+        response_object = {"status": "Fail", "message": str(e)}
+        return make_response(jsonify(response_object)), 400
+
+
+@app.route("/follow-requests", methods=["POST"])
+@authenticate_user
+def follow(**kwargs):
+    try:
+        sender = kwargs.get("current_user")
+        req_data = request.get_json()
+        user = req_data.get("user")
+        receiver = User.query.filter_by(username=user).first()
+        if receiver:
+            if user == sender.username:
+                response_object = {
+                    "status": "Fail",
+                    "message": "You can't send follow request to yourself",
+                }
+                return make_response(jsonify(response_object)), 400
+            elif (
+                Connection.query.filter_by(sender=sender.id, receiver=receiver.id).first()
+                or Connection.query.filter_by(sender=receiver.id, receiver=sender.id).first()
+            ):
+                response_object = {
+                    "status": "Success",
+                    "message": "connection already exists",
+                }
+                return make_response(jsonify(response_object)), 200
+            
+            follow_request = Connection(sender=sender.id, receiver=receiver.id)
+            db.session.add(follow_request)
+            db.session.commit()
+            notification_object = Notification(
+                user=receiver.id,
+                msg="You have a new follow request from {}".format(sender.username),
+            )
+            db.session.add(notification_object)
+            db.session.commit()
+            response_object = {
+                "status": "Success",
+                "message": "Follow request sent to {}".format(receiver.username),
+                "data": connection_schema.dump(follow_request),
+            }
+            return make_response(jsonify(response_object)), 201
+        else:
+            response_object = {
+                "status": "Fail",
+                "message": "unknown username {}".format(user),
+            }
+            return make_response(jsonify(response_object)), 400
+    except Exception as e:
+        response_object = {"status": "Fail", "message": str(e)}
+        return make_response(jsonify(response_object)), 400
+
+
+@app.route("/follow-requests/<id>", methods=["PATCH", "PUT"])
+@authenticate_user
+def respond_to_follow_request(id, **kwargs):
+    try:
+        user = kwargs.get("current_user")
+        follow_request = Connection.query.filter_by(id=id).first()
+        if follow_request.receiver == user.id:
+            if follow_request.accepted == True:
+                response_object = {
+                    "status": "Fail",
+                    "message": "you cant do this, already accepted!",
+                }
+                return make_response(jsonify(response_object)), 400
+            else:
+                if request.get_json()["response"] == "accept":
+                    setattr(follow_request, "accepted", True)
+                    db.session.commit()
+                    notification_object = Notification(
+                        user=follow_request.sender,
+                        msg="{} accepted your follow request ".format(user.username),
+                    )
+                    db.session.add(notification_object)
+                    db.session.commit()
+                    response_object = {"message": "request accepted successfully"}
+                    return make_response(jsonify(response_object)), 200
+                elif request.get_json()["response"] == "reject":
+                    notification_object = Notification(
+                        user=follow_request.sender,
+                        msg="{} rejected your follow request ".format(user.username),
+                    )
+                    db.session.add(notification_object)
+                    db.session.commit()
+                    Connection.query.filter_by(id=id).delete()
+                    db.session.commit()
+                    response_object = {"message": "request deleted successfully"}
+                    return make_response(jsonify(response_object)), 200
+        else:
+            response_object = {
+                "status": "Fail",
+                "message": "you are not authorised for this follow request!",
+            }
+            return make_response(jsonify(response_object)), 400
+    except Exception as e:
+        response_object = {"status": "Fail", "message": str(e)}
+        return make_response(jsonify(response_object)), 400
+
+
+@app.route("/follow-requests/<id>", methods=["DELETE"])
+@authenticate_user
+def unfollow(id, **kwargs):
+    try:
+        user = kwargs.get("current_user")
+        connection = Connection.query.filter_by(id=id).first()
+        if connection:
+            if connection.sender == user.id or connection.receiver == user.id:
+                Connection.query.filter_by(id=id).delete()
+                db.session.commit()
+                response_object = {"status": "success", "message": "connection deleted"}
+                return make_response(jsonify(response_object)), 204
+            else:
+                response_object = {
+                    "status": "Fail",
+                    "message": "you are not authorised!",
+                }
+                return make_response(jsonify(response_object)), 403
+        else:
+            response_object = {"status": "Fail", "message": "no such connection exists!"}
+            return make_response(jsonify(response_object)), 400
+    except Exception as e:
+        response_object = {"status": "Fail", "message": str(e)}
+        return make_response(jsonify(response_object)), 400
+
+
