+++ SocialMediaApplication/views/notification.py
@@ -1,18 +1,19 @@
 from flask import jsonify, make_response
 from flask import current_app as app
-from ..models import db, Notification, notifications_schema
-from .permissions import authenticate_user
+from ..models import db, Notification
+from ..serializers import notifications_schema
+from ..helpers.permissions import authenticate_user
 
 
-@app.route("/notification", methods=["GET"])
+@app.route("/notifications", methods=["GET"])
 @authenticate_user
 def get_all_notifications(**kwargs):
     try:
         user = kwargs.get("current_user")
         if user:
-            notifications = Notification.query.filter_by(user=user.id)
-            notifications_obj = notifications_schema.dump(notifications)
-            responseObject = {"status": "success", "message": notifications_obj}
+            notifications = Notification.query.filter_by(user=user.id).all()
+            notifications_object = notifications_schema.dump(notifications)
+            responseObject = {"status": "success", "message": notifications_object}
             return make_response(jsonify(responseObject)), 200
         else:
             responseObject = {"status": "Fail", "message": "no such user!"}
@@ -22,7 +23,7 @@ def get_all_notifications(**kwargs):
         return make_response(jsonify(responseObject)), 400
 
 
-@app.route("/notification/<id>", methods=["PUT", "PATCH"])
+@app.route("/notifications/<id>", methods=["PUT", "PATCH"])
 @authenticate_user
 def mark_read_notification(id, **kwargs):
     try:
@@ -51,7 +52,7 @@ def mark_read_notification(id, **kwargs):
         return make_response(jsonify(responseObject)), 400
 
 
-@app.route("/notification", methods=["PUT", "PATCH"])
+@app.route("/notifications", methods=["PUT", "PATCH"])
 @authenticate_user
 def mark_all_notifications_as_read(**kwargs):
     try:
@@ -59,7 +60,6 @@ def mark_all_notifications_as_read(**kwargs):
         # print(user)
         if user:
             notifications = Notification.query.filter_by(user=user.id, read=False).all()
-            print(notifications)
             for notification in notifications:
                 if notification.user == user.id:
                     setattr(notification, "read", True)
@@ -77,30 +77,30 @@ def mark_all_notifications_as_read(**kwargs):
         return make_response(jsonify(responseObject)), 400
 
 
-@app.route("/notification/<id>", methods=["DELETE"])
-@authenticate_user
-def delete_read_notification(**kwargs):
-    try:
-        user = kwargs.get("current_user")
-        if user:
-            notification = Notification.query.filter_by(id=id)
-            if notification.user == user.id:
-                notification.delete()
-                db.session.commit()
-                responseObject = {
-                    "status": "success",
-                    "message": "notification deleted!",
-                }
-                return make_response(jsonify(responseObject)), 204
-            else:
-                responseObject = {
-                    "status": "Fail",
-                    "message": "This notification does not belongs to you",
-                }
-                return make_response(jsonify(responseObject)), 400
-        else:
-            responseObject = {"status": "Fail", "message": "no such user!"}
-            return make_response(jsonify(responseObject)), 400
-    except Exception as e:
-        responseObject = {"status": "Fail", "message": str(e)}
-        return make_response(jsonify(responseObject)), 400
+# @app.route("/notifications/<id>", methods=["DELETE"])
+# @authenticate_user
+# def delete_read_notification(**kwargs):
+#     try:
+#         user = kwargs.get("current_user")
+#         if user:
+#             notification = Notification.query.filter_by(id=id)
+#             if notification.user == user.id:
+#                 notification.delete()
+#                 db.session.commit()
+#                 responseObject = {
+#                     "status": "success",
+#                     "message": "notification deleted!",
+#                 }
+#                 return make_response(jsonify(responseObject)), 204
+#             else:
+#                 responseObject = {
+#                     "status": "Fail",
+#                     "message": "This notification does not belongs to you",
+#                 }
+#                 return make_response(jsonify(responseObject)), 400
+#         else:
+#             responseObject = {"status": "Fail", "message": "no such user!"}
+#             return make_response(jsonify(responseObject)), 400
+#     except Exception as e:
+#         responseObject = {"status": "Fail", "message": str(e)}
+#         return make_response(jsonify(responseObject)), 400
