+++ app.py
@@ -1,4 +1,4 @@
-from application import create_app
+from SocialMediaApplication import create_app
 app = create_app()
 
 if __name__ == "__main__":
