--- application/models/__init__.py
@@ -1,5 +1,0 @@
-from .user_model import *
-from .blacklist_token import *
-from .post_model import *
-from .follow_model import *
-from .notification_model import *
