--- application/models/post_model.py
@@ -1,67 +1,0 @@
-from .. import db, ma
-import uuid
-from sqlalchemy.dialects.postgresql import UUID
-from datetime import datetime as dt
-
-
-class PostType(db.Model):
-    __tablename__ = "posttype"
-    id = db.Column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid4,
-        unique=True,
-        nullable=False,
-    )  # Private id
-    name = db.Column(db.String(50), unique=True)
-
-    def __str__(self) -> str:
-        return str(self.name)
-
-
-# Post-type Schema
-class PostTypeSchema(ma.Schema):
-    class Meta:
-        fields = ("name",)
-
-
-post_type_schema = PostTypeSchema()
-post_types_schema = PostTypeSchema(many=True)
-
-
-class Post(db.Model):
-    __tablename__ = "post"
-    id = db.Column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid4,
-        unique=True,
-        nullable=False,
-    )  # Private id
-    creator = db.Column(UUID(as_uuid=True), db.ForeignKey("user.id"))
-    title = db.Column(db.String(50))
-    url = db.Column(db.String(500), default=None)
-    content = db.Column(db.Text)
-    tags = db.Column(db.String(500))
-    pub_date = db.Column(db.DateTime, nullable=False, default=dt.utcnow)
-    post_type = db.Column(db.String(50), db.ForeignKey("posttype.name"))
-    archive = db.Column(db.Boolean, default=False)
-
-
-# Post Schema
-class PostSchema(ma.Schema):
-    class Meta:
-        fields = (
-            "id",
-            "title",
-            "url",
-            "content",
-            "pub_date",
-            "creator",
-            "post_type",
-            "archive",
-        )
-
-
-post_schema = PostSchema()
-posts_schema = PostSchema(many=True)
