--- application/resources/__init__.py
@@ -1,5 +1,0 @@
-from . import users
-from . import post
-from . import search
-from . import follow
-from . import notification
